+++ ChampionsLeagueTickets/ChampionsLeagueTickets.Services/OrderService.cs
@@ -21,22 +21,65 @@ namespace ChampionsLeagueTickets.Services
 
         public async Task AddAsync(Order entity)
         {
-            await _orderDAO.AddAsync(entity);
+            try
+            {
+                if(entity == null)
+                {
+                    throw new ArgumentNullException("Het meegegeven Order mag niet null zijn.");
+                }
+
+                await _orderDAO.AddAsync(entity);
+            }
+            catch(Exception ex)
+            {
+                throw new Exception("Er is iets misgegaan bij het toevoegen van een Order: ", ex);
+            }
         }
 
         public Task DeleteAsync(Order entity)
         {
-            return _orderDAO.DeleteAsync(entity);
+            try
+            {
+                if (entity == null)
+                {
+                    throw new ArgumentNullException("Het meegegeven Order mag niet null zijn.");
+                }
+
+                return _orderDAO.DeleteAsync(entity);
+            }
+            catch (Exception ex)
+            {
+                throw new Exception("Er is iets misgegaan bij het verwijderen van een Order: ", ex);
+            }
         }
 
         public Task<Order?> FindByIdAsync(string Id)
         {
-            return _orderDAO.FindByIdAsync(Id);
+            try
+            {
+                if(Id == null)
+                {
+                    throw new ArgumentNullException("Het meegegeven orderId mag niet null zijn.");
+                }
+
+                return _orderDAO.FindByIdAsync(Id);
+            }
+            catch (Exception ex)
+            {
+                throw new Exception($"Het Order is niet gevonden voor het meegegeven orderId '{Id}': ", ex);
+            }
         }
 
         public async Task<string> GenerateNextOrderIdAsync()
         {
-            return await _orderDAO.GenerateNextOrderIdAsync();
+            try
+            {
+                return await _orderDAO.GenerateNextOrderIdAsync();
+            }
+            catch (Exception ex)
+            {
+                throw new Exception("Het genereren van het volgende orderId is mislukt: ", ex);
+            }
         }
 
         public Task<IEnumerable<Order>?> GetAllAsync()
@@ -44,10 +87,20 @@ namespace ChampionsLeagueTickets.Services
             throw new NotImplementedException();
         }
 
-        public Task<List<Order>> GetAllByUserId(string userId) { 
-   
-           return _orderDAO.GetAllByUserId(userId);
+        public Task<List<Order>> GetAllByUserId(string userId) {
+            try
+            {
+                if(userId == null)
+                {
+                    throw new ArgumentNullException("Het meegegeven userId mag niet null zijn.");
+                }
 
+                return _orderDAO.GetAllByUserId(userId);
+            }
+            catch (Exception ex)
+            {
+                throw new Exception($"Het is mislukt om de orders te vinden voor het userId '{userId}': ", ex);
+            }
         }
     }
 }
